renames e2e test data file to match its (new) contents
